Inspector v2: Handle scene explorer mutations + some bug fixes (#16841)
- If an entity name changes, the associated scene explorer tree item now
updates.
- If an entity parent changes, the change is now reflected immediately
in scene explorer.
- Entities can now be renamed in the properties pane (e.g. it uses a
TextInputPropertyLine rather than a TextPropertyLine).
- useObservableState now updates correctly if the lambda `accessor`
changes. This is needed for cases where the lambda actually changes
because for example a captured value changes. This fixes various things,
including the double mount/unmount that was happening for most cases
where we have an "extensible" list of child components, such as sections
in the property pane.
- Because changing the accessor lambda now causes useObservableState to
re-evaluate, usages of useObservableState now need to memo-ize the
lambda if it is not idempotent (e.g. if each time it returns a different
value, such as a new array or object instance).
- Reduced the number of usages of useObservable by switching various use
cases over to either useProperty or BoundProperty.
- Increased the width of the color picker to make sure the rgba input
fields don't get clipped (since they have a min width).
- Fixed an issue in Playground where it would toggle the inspector twice
when hitting the Play button if the playground code for createScene
returned a promise. This happened to work for the old inspector, but not
for the new inspector (it "correctly" toggles back off).