Inspector v2: Property pane implicit sections, section order, and content order (#16894)
This is something that came up in the recent team push on inspector v2 -
defining property pane sections, order of sections, and order of content
within the sections was confusing to people. As we continue to flesh out
the properties pane, it just adds more and more orders and relationships
between orders, which further increases complexity. With this PR, I'm
making the following changes to address this:

1. Sections do not have to be explicitly registered. If they are not
registered, but referenced when inserting section content, the section
is implicitly created and its implicit order is the order in which it
was first referenced. Sections can still be explicitly registered which
allows extensions to insert a new section in whatever order makes sense.
2. Removed all explicit orders, and instead rely on default ordering
(e.g. order in which the section content was registered).
3. Section identity is now just a string rather than a symbol, which
means as long as the string is the same (e.g. "General", "Transform",
etc.), the content for that section is merged.

I think people will find this easier to work with, and it still allows
for precise and explicit ordering when needed.

In a separate change, I also plan to experiment with removing all of the
*PropertiesService.tsx files and putting all the default properties
section content registrations in one place together within the main
propertiesService.tsx. I hope this be another step in making the
ordering more clear (since it will all be laid out in one file, rather
than spread across files and then order being dependent on the order
those service definitions are added in Inspector.tsx).