Inspector v2: Add useResource hook and use it to correctly manage some object lifetimes in stats (#16917)
One challenge with React function components is managing lifetimes of
disposable resources. I noticed that my previous implementation in stats
was totally wrong and recreating a ton of `SceneInstrumentation` and
`EngineInstrumentation` instances that were never disposed. I have more
cases of this kind of resource management as well, so I'm introducing a
`useResource` custom hook that makes it easy to correctly manage
disposable resources in a React function component, and using this in
stats.

Also, I included one other small fix that prevents registering for the
animation callback for animation properties when the entity has no
animations.