Support vertex pulling (#16826)
Vertex pulling is where you do your own reading of vertex data in the
vertex shader instead of relying on the standard vertex attribute
pipeline. I'm not sure of the best approach in Babylon but I made some
minimal changes to make it work. I'd appreciate some feedback and ideas
for a better way to set this up.

First, some motivation:
I want vertex pulling as a way of selecting vertex data from neighboring
vertices in the IBL Shadows voxelization shader. By retrieving normal
info for the provoking vertex of a triangle, I can selectively swizzle
the axes of the position to maximize rasterization area and avoid
missing voxels. This eliminates the need for 3-pass voxelization and
opens the door for doing realtime voxelization of animated geometry (in
WebGPU only due to the need for 3D storage textures).

https://playground.babylonjs.com/?snapshot=refs/pull/16826/merge#XSNYAU#131

For vertex pulling to work, we typically need:
1. An empty vertex layout.
2. Our needed vertex buffers assigned as storage buffers
3. A non-indexed draw call with the number of vertices to process. This
causes `@builtin(vertex_index)` to be sequential which is critical for
fetching info about neighboring vertices.

No changes are needed to Babylon for the first two requirements. For the
3rd point though, we want to do a `engine.drawArraysType` even though
the mesh has indices. To do this, we set `material.useVertexPulling`.
The shader must also _not_ declare vertex attributes.

---------

Co-authored-by: Popov72 <github@evpopov.com>