Fix GUI 2D optimization and consistently apply states when measuring (#16945)
An old [bug report about TextBlock elements
jittering](https://forum.babylonjs.com/t/glitches-with-textblocks-not-wrapped-in-container/39092)
was fixed by
[disabling](https://github.com/BabylonJS/Babylon.js/commit/b8ce95dd02f7e835f11b4725d7868201f42a7681)
a [prior
optimization](https://github.com/BabylonJS/Babylon.js/pull/13430/files)
to not redraw the 2D control if the new position was the same as the old
position.

That did have the effect of preventing the jitter, but the root cause of
the jitter was that the measurement of the size of the TextBlock
sometimes used the requested font, and sometimes used a default font. If
those were different, the wrong measurement might be used to position
the element, resulting the horizontal shifting back and forth.

With this fix, we consistently apply the requested font before measuring
so the measurement is always correct, and we can now re-enable the perf
optimization to no longer redraw if the projected position of the 2D
control hasn't changed. This addresses the perf concern reported
[here](https://forum.babylonjs.com/t/custom-control-linked-to-a-mesh-is-not-cached-and-redraws-every-frame-dubious-workaround/59822).