Viewer v2: Switch back to WebGL as the default for now (#16946)
Unfortunately we've hit some reliability issues with WebGPU + snapshot
rendering that are difficult to repro and seem to be timing related. We
need to investigate more, but the quick fix to make sure the Viewer is
stable is to switch back to WebGL as the default for now. Might wait for
@Popov72 availability for deeper investigation.

I also added a query param for the Viewer Configurator so that in that
context its easy to test both WebGL and WebGPU.