Exporters: Refactor texture pixel reading (#16972)
This PR aims to unify how pixel data is retrieved in our exporters (and
beyond?). The end result is that both exporters now wait for textures to
be ready, pixels are always returned as Uint8, and the USDZ exporter can
export compressed textures.

- Extend `GetTextureDataAsync` (or optimize it, depending on how you
look at it) such that uncompressed textures are read directly;
compressed textures behave as they did before
- Use just `GetTextureDataAsync` to get pixel data for easier
maintenance, less repetition
- Remove redundant code in glTF exporter
- Add DumpData overloads for QoL
- Add test for texture exports (which, in turn, should be testing
`GetTextureDataAsync` and `DumpData`)

NOTE: The vis test will need to be updated in the future. Right now, we
export compressed textures without their compression, so the
roundtripped texture appears higher quality than before. When we decide
to preserve the compression of textures on export, this test will
probably fail and need updating.