ESLint: Relax naming conventions for destructured vars (#16981)
This change makes it so that if you do something like:

`const { NormalMaterial } = await
import("materials/normal/normalMaterial");`

This is on longer considered a naming convention violation. This applies
to all destructured variables, not just dynamic imports, but since it is
destructuring properties of an object, those properties should already
have been checked for naming conventions (e.g. camelCase).