Inspector v2: Add activate and gizmo commands for cameras (#16995)
This change adds the "Activate and Attach Controls" and "Enable/Disable
Gizmo" commands to Camera nodes in scene explorer.

<img width="1054" height="737" alt="image"
src="https://github.com/user-attachments/assets/3e719425-2005-49cd-b713-a445fd18a800"
/>

This includes a change that adds a `useObservableRenderer` hook, which
builds on `useObservableState` and simply sets a number state to trigger
a re-render. Arguably this is a React anti-pattern, but the alternative
is to have a much more verbose useObservableState that replicates each
property I care about on the command. I'm a bit torn on this and open to
feedback.