Add metadata to Skeleton, MorphTargetManager, and ParticleSystem (#17025)
Currently nodes, materials, textures, and bones come with metadata,
where user can put custom info about the object and can be serialized
with it. But in case of `Skeleton`, `MorphTargetManager`, and
`ParticleSystem`, there does not seem to be a metadata there. This
commit adds a metadata field to `Skeleton`, `MorphTargetManager`, and
`ParticleSystem`, and preserved it in `serialize()`, `Parse()`, and
`clone()`.

Forum post:
<https://forum.babylonjs.com/t/is-there-metadata-for-skeleton/60036>